+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/RecognizePiiEntitiesOptions.java
@@ -13,7 +13,6 @@ import java.util.Arrays;
 @Fluent
 public final class RecognizePiiEntitiesOptions extends TextAnalyticsRequestOptions {
     private PiiEntityDomain domainFilter;
-    private StringIndexType stringIndexType;
     private Iterable<PiiEntityCategory> categoriesFilter;
 
     /**
@@ -106,28 +105,6 @@ public final class RecognizePiiEntitiesOptions extends TextAnalyticsRequestOptio
      */
     public RecognizePiiEntitiesOptions setCategoriesFilter(PiiEntityCategory... categoriesFilter) {
         this.categoriesFilter = Arrays.asList(categoriesFilter);
-        return this;
-    }
-
-    /**
-     * Get the value of {@code stringIndexType}.
-     *
-     * @return The value of {@code stringIndexType}.
-     */
-    public StringIndexType getStringIndexType() {
-        return stringIndexType;
-    }
-
-    /**
-     * Set the value of {@code stringIndexType}.
-     * The {@link StringIndexType#UTF16CODE_UNIT} will be used as default type if there is no value assign to it.
-     *
-     * @param stringIndexType It used to set the value of string indexing type.
-     *
-     * @return the {@link RecognizePiiEntitiesOptions} object itself.
-     */
-    public RecognizePiiEntitiesOptions setStringIndexType(StringIndexType stringIndexType) {
-        this.stringIndexType = stringIndexType;
         return this;
     }
 }